Fix app installation
